(split)bug: 90726 Fixed compilation error.
https://bugzilla.zimbra.com/show_bug.cgi?id=90726

Copied from Perforce
 Change: 541447

+++ src/java/com/zimbra/cs/mailbox/Mailbox.java
@@ -3712,7 +3712,7 @@ public class Mailbox {
         return getModifiedItems(octxt, lastSync, 0, type, folderIds, lastDeleteSync);
     }
 
-    public Pair<List<Integer>, TypedIdList> getModifiedItems(OperationContext octxt, int lastSync, int sinceDate,
+    public Pair<List<Integer>, TypedIdList> getModifiedItems(OperationContext octxt, int lastSync, long sinceDate,
         MailItem.Type type, Set<Integer> folderIds) throws ServiceException {
         return getModifiedItems(octxt, lastSync, sinceDate, type, folderIds, -1);
     }
